Addition of local training feature (#18)
* Initial version for local setup

* Updates to documentation

* Initial version for local setup

* Updates to documentation

* Added cloud detection script

* Azure tweak

* Updates

* Fix bug

* Debugging

* Fix sudo for installing package

* Further bug-fixes

* Fixing wong param in init.

* Additional fix

* Ensuring minio runs

+++ bin/prepare.sh
@@ -1,10 +1,21 @@
 #!/bin/bash
 
+trap ctrl_c INT
+
+function ctrl_c() {
+        echo "Requested to stop."
+        exit 1
+}
+
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
 ## Patch system
 sudo apt-get update && sudo apt-mark hold grub-pc && sudo DEBIAN_FRONTEND=noninteractive apt-get -y -o \
-                        DPkg::options::="--force-confdef" -o DPkg::options::="--force-confold" -qq --force-yes upgrade
+                        DPkg::options::="--force-confdef" -o DPkg::options::="--force-confold" -qq --force-yes upgrade && \
+                        sudo apt-get -y install jq
+
+source $DIR/detect.sh
+echo "Detected cloud type ${CLOUD_NAME}"
 
 ## Do I have a GPU
 GPUS=$(lspci | awk '/NVIDIA/ && /3D controller/' | wc -l)
@@ -16,59 +27,64 @@ fi
 
 ## Do I have an additional disk for Docker images - looking for /dev/sdc (Azure)
 
-ADDL_DISK=$(lsblk | awk  '/^sdc/ {print $1}')
-ADDL_PART=$(lsblk -l | awk -v DISK="$ADDL_DISK" '($0 ~ DISK) && ($0 ~ /part/) {print $1}')
-
-if [ -n $ADDL_DISK ] && [ -z $ADDL_PART];
+if [[ "${CLOUD_NAME}" == "azure" ]];
 then
-    echo "Found $ADDL_DISK, preparing it for use"
-    echo -e "g\nn\np\n1\n\n\nw\n" | sudo fdisk /dev/$ADDL_DISK
-    sleep 1s
-    ADDL_DEVICE=$(echo "/dev/"$ADDL_DISK"1")
-    sudo mkfs.ext4 $ADDL_DEVICE
-    sudo mkdir -p /var/lib/docker
-    echo "$ADDL_DEVICE   /var/lib/docker   ext4    rw,user,auto    0    0" | sudo tee -a /etc/fstab
-    mount /var/lib/docker
-    if [ $? -ne 0 ]
+    ADDL_DISK=$(lsblk | awk  '/^sdc/ {print $1}')
+    ADDL_PART=$(lsblk -l | awk -v DISK="$ADDL_DISK" '($0 ~ DISK) && ($0 ~ /part/) {print $1}')
+
+    if [ -n $ADDL_DISK ] && [ -z $ADDL_PART];
     then
-        echo "Error during preparing of additional disk. Exiting."
-        exit 1
-    fi
-elif  [ -n $ADDL_DISK ] && [ -n $ADDL_PART];
-then
-    echo "Found $ADDL_DISK - $ADDL_PART already mounted. Installing into present drive/directory structure."
+        echo "Found $ADDL_DISK, preparing it for use"
+        echo -e "g\nn\np\n1\n\n\nw\n" | sudo fdisk /dev/$ADDL_DISK
+        sleep 1s
+        ADDL_DEVICE=$(echo "/dev/"$ADDL_DISK"1")
+        sudo mkfs.ext4 $ADDL_DEVICE
+        sudo mkdir -p /var/lib/docker
+        echo "$ADDL_DEVICE   /var/lib/docker   ext4    rw,user,auto    0    0" | sudo tee -a /etc/fstab
+        mount /var/lib/docker
+        if [ $? -ne 0 ]
+        then
+            echo "Error during preparing of additional disk. Exiting."
+            exit 1
+        fi
+    elif  [ -n $ADDL_DISK ] && [ -n $ADDL_PART];
+    then
+        echo "Found $ADDL_DISK - $ADDL_PART already mounted. Installing into present drive/directory structure."
 
-else
-    echo "Did not find $ADDL_DISK. Installing into present drive/directory structure."
+    else
+        echo "Did not find $ADDL_DISK. Installing into present drive/directory structure."
+    fi
 fi
 
-
 ## Do I have an ephemeral disk / temporary storage for runtime output - looking for /dev/nvme0n1 (AWS)?
-
-ADDL_DISK=$(lsblk | awk  '/^nvme0n1/ {print $1}')
-ADDL_PART=$(lsblk -l | awk -v DISK="$ADDL_DISK" '($0 ~ DISK) && ($0 ~ /part/) {print $1}')
-
-if [ -n $ADDL_DISK ] && [ -z $ADDL_PART];
+if [[ "${CLOUD_NAME}" == "aws" ]];
 then
-    echo "Found $ADDL_DISK, preparing it for use"
-    echo -e "g\nn\np\n1\n\n\nw\n" | sudo fdisk /dev/$ADDL_DISK
-    sleep 1s
-    ADDL_DEVICE=$(echo "/dev/"$ADDL_DISK"p1")
-    sudo mkfs.ext4 $ADDL_DEVICE
-    sudo mkdir -p /mnt
-    echo "$ADDL_DEVICE   /mnt   ext4    rw,user,noauto    0    0" | sudo tee -a /etc/fstab
-    mount /mnt
-    if [ $? -ne 0 ]
+
+    ADDL_DISK=$(lsblk | awk  '/^nvme0n1/ {print $1}')
+    ADDL_PART=$(lsblk -l | awk -v DISK="$ADDL_DISK" '($0 ~ DISK) && ($0 ~ /part/) {print $1}')
+
+    if [ -n $ADDL_DISK ] && [ -z $ADDL_PART];
     then
-        echo "Error during preparing of temporary disk. Exiting."
-        exit 1
-    fi
-elif [ -n $ADDL_DISK ] && [ -n $ADDL_PART];
-then
-    echo "Found $ADDL_DISK - $ADDL_PART already mounted, taking no action."
+        echo "Found $ADDL_DISK, preparing it for use"
+        echo -e "g\nn\np\n1\n\n\nw\n" | sudo fdisk /dev/$ADDL_DISK
+        sleep 1s
+        ADDL_DEVICE=$(echo "/dev/"$ADDL_DISK"p1")
+        sudo mkfs.ext4 $ADDL_DEVICE
+        sudo mkdir -p /mnt
+        echo "$ADDL_DEVICE   /mnt   ext4    rw,user,noauto    0    0" | sudo tee -a /etc/fstab
+        mount /mnt
+        if [ $? -ne 0 ]
+        then
+            echo "Error during preparing of temporary disk. Exiting."
+            exit 1
+        fi
+    elif [ -n $ADDL_DISK ] && [ -n $ADDL_PART];
+    then
+        echo "Found $ADDL_DISK - $ADDL_PART already mounted, taking no action."
 
-else
-    echo "Did not find $ADDL_DISK, taking no action."
+    else
+        echo "Did not find $ADDL_DISK, taking no action."
+    fi
 fi
 
 ## Adding Nvidia Drivers
@@ -77,8 +93,8 @@ sudo bash -c 'echo "deb http://developer.download.nvidia.com/compute/cuda/repos/
 sudo bash -c 'echo "deb http://developer.download.nvidia.com/compute/machine-learning/repos/ubuntu1804/x86_64 /" > /etc/apt/sources.list.d/cuda_learn.list'
 sudo bash -c 'apt update && apt install -y nvidia-driver-440 cuda-minimal-build-10-2 -o Dpkg::Options::="--force-overwrite"'
 
-## Adding AWSCli and JQ
-sudo apt-get install -y awscli jq
+## Adding AWSCli
+sudo apt-get install -y awscli 
 
 ## Installing Docker
 curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
@@ -101,9 +117,16 @@ sudo usermod -a -G docker $(id -un)
 sudo curl -L https://github.com/docker/compose/releases/download/1.25.0/docker-compose-`uname -s`-`uname -m` -o /usr/local/bin/docker-compose
 sudo chmod +x /usr/local/bin/docker-compose
 
-## Reboot to load driver -- continue install
-echo "Rebooting in 5 seconds. Will continue with install."
-cd $DIR
-./runonce.sh ./init.sh
-sleep 5s
-sudo reboot
+## Reboot to load driver -- continue install if in cloud-init
+CLOUD_INIT=$(pstree -s $BASHPID | awk /cloud-init/ | wc -l)
+
+if [[ "$CLOUD_INIT" -ne 0 ]];
+then
+    echo "Rebooting in 5 seconds. Will continue with install."
+    cd $DIR
+    ./runonce.sh "./init.sh -m /mnt -c ${CLOUD_NAME}"
+    sleep 5s
+    sudo reboot
+else
+    echo "First stage done. Please reboot and run init.sh"
+fi
